refactor: extract CLI file IO helpers into record-io microcrate
